+++ src/compute-engine/boxed-expression/canonical.ts
@@ -120,8 +120,9 @@ function multiplyForm(expr: BoxedExpression) {
   // If this is a multiply, canonicalize it
   if (expr.head === 'Multiply') return canonicalMultiply(expr.engine, ops);
 
-  if (expr.head === 'Negate')
-    return canonicalMultiply(expr.engine, [expr.op1, expr.engine.NegativeOne]);
+  if (expr.head === 'Negate') {
+    return canonicalMultiply(expr.engine, [ops[0], expr.engine.NegativeOne]);
+  }
 
   return expr;
 }